4.3.5_mod3: 性能优化 - 视频单次读取/帧缓存 + correct_transversalium2 与 fix_edge_effect 向量化 + warpPerspective 替代 skimage + 热循环 np.arange 提到循环外 + logme 句柄缓存 + rot90 k=0 跳过 + cv2_chn.imread dtype 修复

+++ cv2_chn.py
@@ -6,7 +6,8 @@ import numpy as np
 
 def imread(path,arg):
     """读取中文路径图片"""
-    return cv2.imdecode(np.fromfile(path, dtype=np.uint16), -1,arg)
+    # cv2.imdecode expects a 1D byte buffer, so the dtype must be uint8.
+    return cv2.imdecode(np.fromfile(path, dtype=np.uint8), -1,arg)
 
 def imwrite(path, img, arg):
     """保存图片到中文路径"""
